Possible merge.

+++ pytube/models.py
@@ -1,11 +1,14 @@
 from __future__ import unicode_literals
-from os.path import normpath
+from os.path import normpath, isfile
+from os import remove
 from urllib2 import urlopen
+from sys import exit
 
 
 class Video(object):
     """
     Class representation of a single instance of a YouTube video.
+
     """
     def __init__(self, url, filename, **attributes):
         """
@@ -41,32 +44,58 @@ class Video(object):
         """
 
         path = (normpath(path) + '/' if path else '')
-        fullpath = '%s%s.%s' % (path, self.filename, self.extension)
-        response = urlopen(self.url)
-        with open(fullpath, 'wb') as dst_file:
-            meta_data = dict(response.info().items())
-            file_size = int(meta_data.get("Content-Length") or
-                            meta_data.get("content-length"))
-            self._bytes_received = 0
-            while True:
-                self._buffer = response.read(chunk_size)
-                if not self._buffer:
-                    if on_finish:
-                        on_finish(fullpath)
-                    break
+        fullpath = '{0}{1}.{2}'.format(path, self.filename, self.extension)
 
-                self._bytes_received += len(self._buffer)
-                dst_file.write(self._buffer)
-                if on_progress:
-                    on_progress(self._bytes_received, file_size)
+	# Check for conflicting filenames
+	if isfile(fullpath):
+		print "\n\nError: Conflicting filename: '{}'.\n\n".format(self.filename)
+		exit(1)
+
+        response = urlopen(self.url)
+        meta_data = dict(response.info().items())
+        file_size = int(meta_data.get("Content-Length") or
+                         meta_data.get("content-length"))
+        self._bytes_received = 0
+	try:
+        	with open(fullpath, 'wb') as dst_file:
+        # Print downloading message
+       	    		print "\nDownloading: '{0}.{1}' (Bytes: {2})\n\n".format(self.filename, self.extension, file_size)
+
+            		while True:
+                		self._buffer = response.read(chunk_size)
+                		if not self._buffer:
+                    			if on_finish:
+                        			on_finish(fullpath)
+                    			break
+
+                		self._bytes_received += len(self._buffer)
+                		dst_file.write(self._buffer)
+                		if on_progress:
+                    			on_progress(self._bytes_received, file_size)
+
+	# Catch possible exceptions occurring during download
+	except IOError:
+		print """\n\nError: Failed to open file.\nCheck that: ('{0}'), is a valid pathname.
+			 	\nOr that ('{1}.{2}') is a valid filename.\n\n""".format(path, self.filename, self.extension)
+		exit(2)
+
+	except BufferError:
+		print "\n\nError: Failed on writing buffer.\nFailed to write video to file.\n\n"
+		exit(1)
+
+	except KeyboardInterrupt:
+		print "\n\nInterrupt signal given.\nDeleting incomplete video ('{0}.{1}').\n\n".format(self.filename, self.extension)
+		remove(fullpath)
+		exit(1)
+
 
     def __repr__(self):
         """A cleaner representation of the class instance."""
-        return "<Video: %s (.%s) - %s>" % (self.video_codec, self.extension,
+        return "<Video: {0} (.{1}) - {2}>".format(self.video_codec, self.extension,
                                            self.resolution)
 
     def __lt__(self, other):
         if type(other) == Video:
-            v1 = "%s %s" % (self.extension, self.resolution)
-            v2 = "%s %s" % (other.extension, other.resolution)
+            v1 = "{0} {1}".format(self.extension, self.resolution)
+            v2 = "{0} {1}".format(other.extension, other.resolution)
             return (v1 > v2) - (v1 < v2) < 0
